+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - BusinessAccountRequest resource
 - validator_link attribute to IndividualAccountRequest resource
 - complement attribute to IndividualAccountRequest address
+### Fixed
+- business-rule and enum drift in docstrings across Pix, Issuing and Credit resources
 
 ## [0.28.0] - 2026-06-24
 ### Added